requirements.txt
python-telegram-bot==20.0a2
pygraphic==0.3
requests==2.28.1
